人事管理系统 有三种访问权限： 0——可以访问系统所有功能（总经理） 1——可以访问员工系统和用户系统的所有功能（人事部） 2——只能访问员工系统、用户系统和公告系统的部分功能，即对本人员工信息进行查询和部分修改、对本人用户名、密码进行查询和修改、公告系统查询。 共有5个模块： 员工模块 用户模块 部门模块 职位模块 公告模块 这5个模块都有自己的增删改查功能

+++ src/main/java/java_personnel_system/service/DepartmentService.java
@@ -35,9 +35,15 @@ public class DepartmentService {
             Print.print("部门不存在，请重新输入");
             removeDepartmentId();
         }
-        staffDao.autoUpdateStaffDepartmentId(departmentId);
-        departmentDao.removeDepartmentId(departmentId);
-        departmentDao.departmentAlter();
+        Print.print("该部门还有员工，请确认是否删除");
+        Print.print("1.是");
+        Print.print("2.否");
+        int confirm = sc.nextInt();
+        if (confirm==1){
+            staffDao.autoUpdateStaffDepartmentId(departmentId);
+            departmentDao.removeDepartmentId(departmentId);
+            departmentDao.departmentAlter();
+        }
     }
 
     public static void removeDepartmentName() throws Exception{
@@ -47,7 +53,7 @@ public class DepartmentService {
             Print.print("部门不存在，请重新输入");
             removeDepartmentName();
         }
-        Print.print("该部门还要员工，请确认是否删除");
+        Print.print("该部门还有员工，请确认是否删除");
         Print.print("1.是");
         Print.print("2.否");
         int confirm = sc.nextInt();
